requirements.txt
Flask==3.1.0
SQLAlchemy==2.0.37
APScheduler==3.10.4
requests==2.32.3
urllib3<2
beautifulsoup4==4.12.3
python-dateutil==2.9.0.post0
gunicorn==23.0.0
